+++ test/ComparedQueryable.Test/NativeQueryableTests/AnyTests.cs
@@ -2,11 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+using System.Linq;
 using System.Linq.Expressions;
-using ComparedQueryable.Test.NativeQueryableTests;
 using Xunit;
 
-namespace System.Linq.Tests
+namespace ComparedQueryable.Test.NativeQueryableTests
 {
     public class AnyTests : EnumerableBasedTests
     {
