+++ test/ComparedQueryable.Test/NativeQueryableTests/EnumerableTests.cs
@@ -3,8 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using System.Linq;
 
-namespace System.Linq.Tests
+namespace ComparedQueryable.Test.NativeQueryableTests
 {
     public abstract class EnumerableBasedTests
     {
