+++ test/ComparedQueryable.Test/NativeQueryableTests/IntersectTests.cs
@@ -2,11 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Generic;
-using ComparedQueryable.Test.NativeQueryableTests;
+using System.Linq;
 using Xunit;
 
-namespace System.Linq.Tests
+namespace ComparedQueryable.Test.NativeQueryableTests
 {
     public class IntersectTests : EnumerableBasedTests
     {
