+++ test/ComparedQueryable.Test/NativeQueryableTests/ReverseTests.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using ComparedQueryable.Test.NativeQueryableTests;
+using System;
+using System.Linq;
 using Xunit;
 
-namespace System.Linq.Tests
+namespace ComparedQueryable.Test.NativeQueryableTests
 {
     public class ReverseTests : EnumerableBasedTests
     {
